refactor project Part 2
