--- order-service/src/main/java/com/ecommerce/orderservice/repository/OrderRepository.java
@@ -1,13 +1,0 @@
-package com.ecommerce.orderservice.repository;
-
-import com.ecommerce.orderservice.model.entity.OrderEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface OrderRepository extends JpaRepository<OrderEntity, UUID> {
-
-}
-
